==> 11-Storytelling.R <==
####  Storytelling  ####

# Configurando o diretório de trabalho
setwd("~/Desktop/DataScience/CienciaDeDados/1.Big-Data-Analytics-com-R-e-Microsoft-Azure-Machine-Learning/14.Projeto-de-Analise-de-Regressao-com-R-e-Azure-ML-Prevendo_Demanda_Aluguel_Bicicleta")
getwd()

 

# -> Todo Cientista de Dados precisa desenvolver a habilidade de conseguitar contar a história do seu projeto


# -> Como resumir todo um trabalho de análise e explicar isso a quem não conhece Data Science, Estatística, Programação ou
#    Machine Learning ? Aplicando a técnica de Storytelling.



## Dicas para Storytelling

# - Não utilizar linguagem técnica em apresentações executivas.

# - Use apenas um ou dois gráficos para explicar seus resultados (use tabelas se necessário).

# - Tente construir uma linha do tempo, explicando como os dados geraram o resultado preditivo esperado.

# - Não use termos como "variáveis" e sim "atributos" ou "características".

# - Documente todo seu trabalho, pois isso servirá de apoio para que você possa contar a história dos dados.

# - Os tomadores de decisão não estão interessados em saber se você usou um algoritmo de regressão ou de classificação e nem
#   como você otimizou seu modelo. Eles querem saber o que precisam fazer para aumentar as vendas e o faturamento!

# - A fase de Análise Exploratóra dos Dados podem apresentar insights preciosos. Use isto a seu favor e guarde todos os resultados
#   intermediários do seu processo de análise!
